+++ platform/windows/cs/Adguard.Dns/Adguard.Dns/Api/DnsProxyServer/Configs/DnsStamp.cs
@@ -5,6 +5,7 @@ namespace Adguard.Dns.Api.DnsProxyServer.Configs
 {
     /// <summary>
     /// DNS Stamp
+    /// (a managed mirror of <see cref="AGDnsApi.ag_dns_stamp"/>)
     /// </summary>
     public class DnsStamp
     {
@@ -45,6 +46,18 @@ namespace Adguard.Dns.Api.DnsProxyServer.Configs
         /// </summary>
         public AGDnsApi.ag_server_informal_properties Properties { get; set; }
 
+        /// <summary>
+        /// A URL representation of this stamp which can be used
+        /// as a valid ag_upstream_options address
+        /// </summary>
+        public string PrettyUrl { get; set; }
+
+        /// <summary>
+        /// A URL representation of this stamp which is prettier,
+        /// but can NOT be a valid ag_upstream_options address
+        /// </summary>
+        public string PrettierUrl { get; set; }
+
         #region Equals members
 
         public override bool Equals(object obj)
@@ -75,7 +88,9 @@ namespace Adguard.Dns.Api.DnsProxyServer.Configs
                    DoHPath == other.DoHPath &&
                    CollectionUtils.ListsEquals(new List<byte>(PublicKey), new List<byte>(other.PublicKey)) &&
                    Hashes.Count == other.Hashes.Count &&
-                   Properties == other.Properties;
+                   Properties == other.Properties &&
+                   PrettyUrl == other.PrettyUrl &&
+                   PrettierUrl == other.PrettierUrl;
         }
 
         public override int GetHashCode()
@@ -89,6 +104,8 @@ namespace Adguard.Dns.Api.DnsProxyServer.Configs
                 hashCode = (hashCode * 397) ^ (DoHPath != null ? PublicKey.Length : 0);
                 hashCode = (hashCode * 397) ^ (DoHPath != null ? Hashes.Count : 0);
                 hashCode = (hashCode * 397) ^ Properties.GetHashCode();
+                hashCode = (hashCode * 397) ^ (PrettyUrl != null ? PrettyUrl.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ (PrettierUrl != null ? PrettierUrl.GetHashCode() : 0);
                 return hashCode;
             }
         }